\name{spray_missing_accessor}
\alias{spray_missing_accessor}
\title{Discussion document}
\description{
Discussion about the difficulties of implementing idiom like
\code{S[1,,5,,]} in the package
}
\usage{
spray_missing_accessor(S, dots)
}
\arguments{
  \item{S}{Object of class spray}
  \item{dots}{further arguments}
}
\details{
Look at the source which contains an extended discussion of the
difficulties
}
\author{Robin K. S. Hankin}
