\name{spray-class}
\docType{class}
\alias{spray-class}
\title{Class \dQuote{spray}}
\description{The formal S4 class for sprays.}
\section{Objects from the Class}{

  A spray object is typically a list with components \code{index} and
  \code{value}.
  
}
\author{Robin K. S. Hankin}
\seealso{\code{\link{spray}}}
\keyword{classes}
